3D: keep span attributes in the width the pixel stages read
The span carried five s32 planes for r g b s t, but the shader reads colour
as a 6-bit channel and the sampler truncates coordinates to s16. span_attrs5n
narrows them as it stores: three u8 planes and two s16 planes, eight pixels a
step instead of four, and the texture gathers take s16 coordinates directly
instead of shifting them out of 32-bit words.

SpanBuf drops from ~7.4 KB to ~3.5 KB.

Exact: kernel twins (new span_attrs5n test), the 8 host baselines, NEON frame
dumps under qemu for seven games.

+++ src/core/gpu/render3d.h
@@ -156,7 +156,7 @@ private:
   // Four pixels per step; same results as resolve_span (the specification),
   // for polygons without shadow / wireframe / toon shading.
   template <int mode, bool textured, bool aa> void resolve_span_vec(const Shade& sh, const SpanBuf& sb, s32 y, s32 xa, s32 xb, int part, int edge, s32 l_cov, s32 r_cov, s32& xcov);
-  void texture_gather4(const Shade& sh, const s32* sa, const s32* ta, u32* colour, u32* alpha) const;
+  void texture_gather4(const Shade& sh, const s16* sa, const s16* ta, u32* colour, u32* alpha) const;
   // A textured span's texels, gathered once into the span buffer before the
   // resolve loop reads them (removes an indirect call per four pixels).
   void span_texels(const Shade& sh, SpanBuf& sb, s32 ca, s32 cb) const;
